refactor: move interfaces to separate package common
