+++ internal/common/igame.go
@@ -1,0 +1,7 @@
+package common
+
+type IGame interface {
+	GetWinnerString() (string, error)
+	NextMove(int, int) error
+	ToHtml(...string) string
+}
