+++ internal/common/iplayer.go
@@ -1,0 +1,5 @@
+package common
+
+type IPlayer interface {
+	GetMark() rune
+}
